reorganize torin
