+++ common/src/layout.rs
@@ -1,5 +1,5 @@
 use std::ops::Div;
-use torin::{Area, Size2D};
+use torin::geometry::{Area, Size2D};
 
 /// Layout info of a certain Node, used by `use_node`.
 #[derive(Clone, Debug, Default, PartialEq)]
