+++ freya/src/lib.rs
@@ -83,7 +83,7 @@ pub mod prelude {
     pub use freya_hooks::*;
     pub use freya_node_state::{bytes_to_data, CustomAttributeValues};
     pub use freya_renderer::WindowConfig;
-    pub use torin;
+    pub use torin::prelude::*;
 
     pub use tracing;
 }
