+++ renderer/src/elements/rect_container.rs
@@ -2,7 +2,7 @@ use dioxus_native_core::real_dom::NodeImmutable;
 use freya_dom::DioxusNode;
 use freya_node_state::{References, Style};
 use skia_safe::{BlurStyle, Canvas, MaskFilter, Paint, PaintStyle, Path, PathDirection, Rect};
-use torin::Area;
+use torin::geometry::Area;
 
 /// Render a `rect` or a `container` element
 pub fn render_rect_container(area: &Area, node_ref: &DioxusNode, canvas: &mut Canvas) {
