+++ renderer/src/lib.rs
@@ -9,7 +9,7 @@ use freya_dom::SafeDOM;
 use freya_elements::events::keyboard::{
     from_winit_to_code, get_modifiers, get_non_text_keys, Code, Key,
 };
-use torin::CursorPoint;
+use torin::geometry::CursorPoint;
 
 use std::sync::{Arc, Mutex};
 use winit::event::{
