+++ torin/src/geometry.rs
@@ -1,0 +1,8 @@
+#[derive(PartialEq)]
+pub struct Measure;
+
+pub type Area = euclid::Rect<f32, Measure>;
+pub type Size2D = euclid::Size2D<f32, Measure>;
+pub type Point2D = euclid::Point2D<f32, Measure>;
+pub type CursorPoint = euclid::Point2D<f64, Measure>;
+pub type Length = euclid::Length<f32, Measure>;
